refactor: checkbox 리팩터링
